fix: pass deltas to handler one by one
If we pass the whole batch to the handler and it throws for whatever
reason, we will lose the entire batch. We could add error handling
to our own handler, but then we also expect any other project using
this service and implementing their own handler to also implement
error handling on their own level.

Instead, we handle errors at the service side and ensure the handler
only ever gets one delta.

In the future, if other projects start using this service, this might prove
to be undesired, in that case the change is pretty simple, revert this
commit and implement error handling on our own handler and make
it part of the service's "contract" to have the delta handler deal with errors.